update to 10.13.1 (4845)

+++ TMessagesProj/src/main/java/org/telegram/ui/Charts/view_data/ChartHorizontalLinesData.java
@@ -43,7 +43,7 @@ public class ChartHorizontalLinesData {
                 v = round(newMaxHeight);
             }
 
-            int step = Math.max(1, (int) Math.ceil(v / 5.0));
+            long step = Math.max(1, (long) Math.ceil(v / 5.0));
 
             int n;
             if (v < 6) {
@@ -71,7 +71,7 @@ public class ChartHorizontalLinesData {
                 if (k > 0) {
                     float v2 = (values[i] / k);
                     if (skipFloatValues) {
-                        if (v2 - ((int) v2) < 0.01f || formatter == ChartData.FORMATTER_TON) {
+                        if (v2 - ((long) v2) < 0.01f || formatter == ChartData.FORMATTER_TON) {
                             valuesStr2[i] = format(1, secondTextPaint, (long) v2, formatter);
                         } else {
                             valuesStr2[i] = "";
@@ -114,12 +114,12 @@ public class ChartHorizontalLinesData {
             }
             boolean skipFloatValues = step / k < 1;
             for (int i = 0; i < n; i++) {
-                values[i] = newMinHeight + (int) (i * step);
+                values[i] = newMinHeight + (long) (i * step);
                 valuesStr[i] = format(0, firstTextPaint, newMinHeight + (long) (i * step), formatter);
                 if (k > 0) {
                     float v = (values[i] / k);
                     if (skipFloatValues) {
-                        if (v - ((int) v) < 0.01f || formatter == ChartData.FORMATTER_TON) {
+                        if (v - ((long) v) < 0.01f || formatter == ChartData.FORMATTER_TON) {
                             valuesStr2[i] = format(1, secondTextPaint, (long) v, formatter);
                         } else {
                             valuesStr2[i] = "";
@@ -152,13 +152,13 @@ public class ChartHorizontalLinesData {
         return AndroidUtilities.formatWholeNumber((int) v, 0);
     }
 
-    public static int lookupHeight(long maxValue) {
+    public static long lookupHeight(long maxValue) {
         long v = maxValue;
         if (maxValue > 100) {
             v = round(maxValue);
         }
 
-        int step = (int) Math.ceil(v / 5f);
+        long step = (long) Math.ceil(v / 5f);
         return step * 5;
     }
